Refactor wehook faber and alice
