+++ src/main/java/webhook/alice/GlobalService.java
@@ -14,7 +14,7 @@ import java.util.logging.Logger;
 import static utils.Common.prettyJson;
 
 @Service
-public class InitService {
+public class GlobalService {
     // get logger for demo - INFO configured
     static final Logger logger = Common.getDemoLogger();
 
@@ -27,7 +27,7 @@ public class InitService {
 
         // static configuration
         long utime = System.currentTimeMillis() / 1000;
-        DocumentContext provisionConfig = JsonPath.parse("{" +
+        String provisionConfig = JsonPath.parse("{" +
                 //"  agency_url: 'http://15.165.161.165:8080'," + // skt test dummy cloud agent
                 "  agency_url: 'http://localhost:8080'," + // use local
                 "  agency_did: 'VsKV7grR1BUE29mG2Fm2kX'," +
@@ -36,28 +36,28 @@ public class InitService {
                 "  wallet_key: '123'," +
                 "  payment_method: 'null'," +
                 "  enterprise_seed: '000000000000000000000000000User1'" + // SEED of alice's DID that does not need to be registered in the ledger
-                "}");
+                "}").jsonString();
 
         // Communication method. aries.
-        provisionConfig.put("$", "protocol_type", "3.0");
+        provisionConfig = JsonPath.parse(provisionConfig).put("$", "protocol_type", "3.0").jsonString();
         logger.info("Running with Aries VCX Enabled! Make sure VCX agency is configured to use protocol_type 3.0");
 
         // add webhook url to config
-        provisionConfig.put("$", "webhook_url", webhookUrl);
+        provisionConfig = JsonPath.parse(provisionConfig).put("$", "webhook_url", webhookUrl).jsonString();
         logger.info("Running with webhook notifications enabled! Webhook url = " + webhookUrl);
 
-        logger.info("#8 Provision an agent and wallet, get back configuration details: \n" + prettyJson(provisionConfig.jsonString()));
-        DocumentContext vcxConfig = JsonPath.parse(UtilsApi.vcxProvisionAgent(provisionConfig.jsonString()));
+        logger.info("#8 Provision an agent and wallet, get back configuration details: \n" + prettyJson(provisionConfig));
+        String vcxConfig = UtilsApi.vcxProvisionAgent(provisionConfig);
 
-        vcxConfig.put("$", "institution_name", "alice")
+        vcxConfig = JsonPath.parse(vcxConfig).put("$", "institution_name", "alice")
                 .put("$", "institution_logo_url", "http://robohash.org/345")
                 .put("$", "protocol_version", "2")
-                .put("$", "genesis_path", System.getProperty("user.dir") + "/genesis.txn"); // file configured for skt testnet
-        //.put("$", "genesis_path", "http://54.180.86.51/genesis"); // or url can be configured
-        logger.info("#9 Initialize libvcx with new configuration\n" + prettyJson(vcxConfig.jsonString()));
-        VcxApi.vcxInitWithConfig(vcxConfig.jsonString()).get();
+                .put("$", "genesis_path", System.getProperty("user.dir") + "/genesis.txn").jsonString(); // file configured for skt testnet
+        //.put("$", "genesis_path", "http://54.180.86.51/genesis").jsonString(); // or url can be configured
+        logger.info("#9 Initialize libvcx with new configuration\n" + prettyJson(vcxConfig));
+        VcxApi.vcxInitWithConfig(vcxConfig).get();
 
-        WalletApi.addRecordWallet("vcxConfig", "defaultVcxConfig", vcxConfig.jsonString()).get();
+        WalletApi.addRecordWallet("vcxConfig", "defaultVcxConfig", vcxConfig).get();
 
         // TODO: may vcxUpdateWebhookUrl is called during vcxInitWithConfig
         VcxApi.vcxUpdateWebhookUrl(webhookUrl).get();
